Ignore package-lock.json files
We are using `yarn` in this project and are intentionally checking-in
`yarn.lock`. Some people seem to prefer using `npm` locally, but we don't want
the `package-lock.json` file to also be checked in. It's probably easier to just
ignore this file than to have this discussion in the respective pull requests :)

Example: https://github.com/hyper-scale/race/pull/93